Simulated Data of Student Test Scores

Simulated Data of Student Test Scores

Usage

data("data_student")

Format

A data frame with 41 observations on the following 4 variables.

student

student ID

class

class/subject

test1

score on test 1

test2

score on test 2

Details

This simulated data imagines the following scenario:

In a school, 12 students (A-M) were tested in 6 classes (math, chemistry, physics, art, horticulture, welding).

Not all students were enrolled in all classes.

In each class, students were given test 1, then later test 2. Test scores could be 0-100.

(Using 2 tests is typical for learning assessment, measuring intervention effectiveness, comparison of test forms, etc.)

Some students missed class on the day of a test and had no score for that test.

Student B felt ill during art test 1 and was allowed to re-take test 1 (both test 1 scores are included and the same test 2 score was used).

Source

None

References

None

Examples

data(data_student)